+++ sdk/translation/azure-ai-translation-document/azure/ai/translation/document/_operations/_patch.py
@@ -81,7 +81,7 @@ def convert_status(status, ll=False):
 
 class DocumentTranslationLROPoller(LROPoller[PollingReturnType_co]):
     """A custom poller implementation for Document Translation. Call `result()` on the poller to return
-    a pageable of :class:`~azure.ai.translation.document.DocumentStatus`."""
+    a pageable of :class:`~azure.ai.translation.document.models.DocumentStatus`."""
 
     _polling_method: "DocumentTranslationLROPollingMethod"
 
@@ -102,7 +102,7 @@ class DocumentTranslationLROPoller(LROPoller[PollingReturnType_co]):
         """The details for the translation operation
 
         :return: The details for the translation operation.
-        :rtype: ~azure.ai.translation.document.TranslationStatus
+        :rtype: ~azure.ai.translation.document.models.TranslationStatus
         """
         # pylint: disable=protected-access
         if self._polling_method._current_body: